+++ apps/api-harmonization/src/app.module.ts
@@ -24,6 +24,7 @@ import {
 import { configuration } from '@o2s/api-harmonization/config/configuration';
 
 import { ArticleListBlockModule } from '@o2s/api-harmonization/blocks/article-list/article-list.module';
+import { ArticleSearchBlockModule } from '@o2s/api-harmonization/blocks/article-search/article-search.module';
 import { ArticleBlockModule } from '@o2s/api-harmonization/blocks/article/article.module';
 import { CategoryListBlockModule } from '@o2s/api-harmonization/blocks/category-list/category-list.module';
 import { CategoryBlockModule } from '@o2s/api-harmonization/blocks/category/category.module';
@@ -109,6 +110,7 @@ import { SurveyjsModule } from './modules/surveyjs-forms/surveyjs.module';
         ArticleListBlockModule.register(AppConfig),
         CategoryBlockModule.register(AppConfig),
         ArticleBlockModule.register(AppConfig),
+        ArticleSearchBlockModule.register(AppConfig),
         // BLOCK REGISTER
     ],
     providers: [