+++ apps/frontend/next.config.ts
@@ -25,6 +25,10 @@ const nextConfig: NextConfig = {
                 protocol: 'https',
                 hostname: 'picsum.photos',
             },
+            {
+                protocol: 'https',
+                hostname: 'medusa-public-images.s3.eu-west-1.amazonaws.com',
+            },
         ],
     },
     sassOptions: {