+++ apps/frontend/src/components/ContentSection/ContentSection.types.ts
@@ -1,9 +1,8 @@
+import { Models } from '@o2s/framework/modules';
+
 export interface ContentSectionProps {
     title?: string;
     description?: string;
-    additionalLink?: {
-        url: string;
-        label: string;
-    };
+    categoryLink?: Models.Link.Link;
     children: React.ReactNode;
 }